pyproject.toml
[project]
name = "aaib-collector"
version = "0.1.0"
description = "Automated AAIB report collector and structured data extractor"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fitz>=0.0.1.dev2",
    "requests>=2.32.5",
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",
    "openai>=1.0.0",
    "pymupdf>=1.24.0",
    "python-dotenv>=1.0.0",
]
